fix(webapp): hide webapp hash behind vanity url
using the compiled app delivered as a hash causes sad emoji when the api
is up in read-only mode b/c then the user can't access the /ipfs endpoint
on the api, so to fix this I've cleaned up our webapp code a bit in the api
package and now the frontend just loads "webapp.js", with new logic in the
backend that knows what to do with that.
