+++ config/webapp.go
@@ -9,21 +9,18 @@ type Webapp struct {
 	Port    string
 	// token for analytics tracking
 	AnalyticsToken string
-	// Scripts is a list of script tags to include the webapp page, useful for using alternative / beta
-	// versions of the qri frontend
-	Scripts []string
+	// EntrypointHash is a hash of the compiled webapp (the output of running webpack https://github.com/qri-io/frontend)
+	// this is fetched and replaced via dnslink when the webapp server starts
+	// the value provided here is just a sensible fallback for when dnslink lookup fails,
+	// pointing to a known prior version of the the webapp
+	EntrypointHash string
 }
 
 // Default creates a new default Webapp configuration
 func (Webapp) Default() *Webapp {
 	return &Webapp{
-		Enabled: true,
-		Port:    DefaultWebappPort,
-		Scripts: []string{
-			// this is fetched and replaced via dnslink when the webapp server starts
-			// the value provided here is just a sensible fallback for when dnslink lookup fails,
-			// pointing to a known prior version of the the webapp
-			"http://localhost:2503/ipfs/QmewSfmnridhYdwLc9nGVwFNhMofAjPf4vxMMUYC6QDEjm",
-		},
+		Enabled:        true,
+		Port:           DefaultWebappPort,
+		EntrypointHash: "QmewSfmnridhYdwLc9nGVwFNhMofAjPf4vxMMUYC6QDEjm",
 	}
 }
